changed video stream flow to begin at frontend
Instead of the previous backend video capture being sent to frontend, the frontend is now deeply integrated in the flow of video. Video starts in the frontend, is processed on the backend, and is then sent back to the frontend all over the same websocket. Including selection of processing types.